+++ ChangeLog.md
@@ -27,8 +27,8 @@ handling for several very different languages (proofs-of-concept).
 
 First hackage release
 
- - [ ] Haddock documentation
- - [.] Update README
+ * Haddock documentation
+ * Update README
  * In help statements write universal, not concrete, types
 
 0.32.0 [2020.11.01]